refactor: Extract app setup logic refactor: Extract flask form logic
